Bug 1675602 - Move other ToWebAssemblyValue/ToJSValue implementations to WasmTypes.h. r=lth
There are two independent implementations of ToWebAssemblyValue/
ToJSValue. One in WasmJS.cpp and another in WasmInstance.cpp. The
first was moved to WasmTypes.cpp in the previous commit. This commit
moves the second over to WasmTypes.cpp as well. A later commit
will unify the two implementations.

Differential Revision: https://phabricator.services.mozilla.com/D96219
